@@ -1,5 +1,5 @@
 import {Component, OnChanges, OnInit, SimpleChanges} from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
 
 import { DefaultFilter } from 'ng2-smart-table';
@@ -16,7 +16,7 @@ import { DefaultFilter } from 'ng2-smart-table';
   `,
 })
 export class CustomFilterComponent extends DefaultFilter implements OnInit, OnChanges {
-  inputControl = new FormControl();
+  inputControl = new UntypedFormControl();
 
   constructor() {
     super();
