@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 import { DefaultFilter } from './default-filter';
 import { debounceTime } from 'rxjs/operators';
@@ -15,7 +15,7 @@ import { debounceTime } from 'rxjs/operators';
 export class CheckboxFilterComponent extends DefaultFilter implements OnInit {
 
   filterActive: boolean = false;
-  inputControl = new FormControl();
+  inputControl = new UntypedFormControl();
 
   constructor() {
     super();
